+++ packages/annotations/src/mutator.ts
@@ -1,5 +1,5 @@
 import { getParameterNames } from '@alterior/common';
-import { DecoratorSite, AnnotationDecoratorOptions, Annotation, MutatorDefinition, AnnotationDecorator } from './annotations';
+import { DecoratorSite, AnnotationDecoratorOptions, Annotation, AnnotationDecorator, ClassDecoratorSite, MethodDecoratorSite, ParameterDecoratorSite, PropertyDecoratorSite } from './annotations';
 
 /**
  * Mutators are a way to define "mutation decorators" which in some way change the value
@@ -9,19 +9,47 @@ import { DecoratorSite, AnnotationDecoratorOptions, Annotation, MutatorDefinitio
  * Create a mutator with Mutator.create().
  */
 
+type MutatorFunction<SiteType extends DecoratorSite = DecoratorSite, Args extends any[] = any[]> = 
+    (target: SiteType, ...args: Args) => void;
+
+export interface MutatorDefinition<SiteType extends DecoratorSite = DecoratorSite, Args extends any[] = any[]> {
+    invoke: (site: SiteType, ...args: Args) => void;
+    options?: AnnotationDecoratorOptions<void>;
+}
+
 export class Mutator {
     /**
      * Low-level method to ceate a new mutation decorator (mutator) based on the given function.
      * Use `Mutator.define()` instead.
      */
     public static create<Args extends any[]>(
-        mutator: (target: DecoratorSite, ...args: Args) => void, 
+        mutator: MutatorFunction<ClassDecoratorSite, Args>, 
+        options?: AnnotationDecoratorOptions<void> & { validTargets: ['class'] }
+    ): AnnotationDecorator<Args>;
+    public static create<Args extends any[]>(
+        mutator: MutatorFunction<MethodDecoratorSite, Args>, 
+        options?: AnnotationDecoratorOptions<void> & { validTargets: ['method'] }
+    ): AnnotationDecorator<Args>;
+    public static create<Args extends any[]>(
+        mutator: MutatorFunction<PropertyDecoratorSite, Args>, 
+        options?: AnnotationDecoratorOptions<void> & { validTargets: ['property'] }
+    ): AnnotationDecorator<Args>;
+    public static create<Args extends any[]>(
+        mutator: MutatorFunction<ParameterDecoratorSite, Args>, 
+        options?: AnnotationDecoratorOptions<void> & { validTargets: ['parameter'] }
+    ): AnnotationDecorator<Args>;
+    public static create<Args extends any[]>(
+        mutator: MutatorFunction<DecoratorSite, Args>, 
+        options?: AnnotationDecoratorOptions<void>
+    ): AnnotationDecorator<Args>;
+    public static create<Args extends any[]>(
+        mutator: MutatorFunction<any, Args>, 
         options?: AnnotationDecoratorOptions<void>
     ): AnnotationDecorator<Args> {
         return <AnnotationDecorator<any[]>> Annotation.decorator(Object.assign({}, options || {}, {
             factory: (target: DecoratorSite, ...args: Args) => {
                 let paramNames: string[] | undefined;
-                let value = target.propertyDescriptor?.value;
+                let value = (target as any).propertyDescriptor?.value;
 
                 if (typeof value === 'function') {
                     paramNames = getParameterNames(value);
@@ -29,7 +57,7 @@ export class Mutator {
 
                 mutator(target, ...args);
 
-                let replacement = target.propertyDescriptor?.value;
+                let replacement = (target as any).propertyDescriptor?.value;
                 if (value !== replacement && paramNames !== undefined && !Object.hasOwn(replacement, '__parameterNames')) {
                     Object.defineProperty(replacement, '__parameterNames', {
                         value: paramNames
@@ -44,15 +72,15 @@ export class Mutator {
      * This should be called and returned from a
      * function definition. For example:
      *
-```
-function Name() {
-    return Mutator.define({
-        invoke(site) {
-            // ...
-        }
-    })
-}
-```
+     * ```
+     * function Name() {
+     *     return Mutator.define({
+     *         invoke(site) {
+     *             // ...
+     *         }
+     *     })
+     * }
+     * ```
      *
      * The `invoke()` function takes a DecoratorSite object which describes the particular
      * invocation that is being run, and importantly, access to the property descriptor
@@ -62,21 +90,27 @@ function Name() {
      *
      * For example:
      * ```
-export function RunTwice() {
-  return Mutator.create(
-    site => {
-      let prop = site.propertyDescriptor;
-      let original = prop.value;
-      let replacement = function(...args) {
-        original.apply(this, args);
-        original.apply(this, args);
-      }
-      prop.value = replacement;
-    }
-)
+     * export function RunTwice() {
+     *     return Mutator.define(
+     *         invoke(site) {
+     *             let prop = site.propertyDescriptor;
+     *             let original = prop.value;
+     *             let replacement = function(...args) {
+     *                 original.apply(this, args);
+     *                 original.apply(this, args);
+     *             }
+     *             prop.value = replacement;
+     *         }
+     *     );
+     * }
      * ```
      */
-    public static define(definition: MutatorDefinition) {
+    public static define<Args extends any[]>(definition: MutatorDefinition<ClassDecoratorSite, Args> & { options: { validTargets: ['class'] }}): AnnotationDecorator<Args>;
+    public static define<Args extends any[]>(definition: MutatorDefinition<MethodDecoratorSite, Args> & { options: { validTargets: ['method'] }}): AnnotationDecorator<Args>;
+    public static define<Args extends any[]>(definition: MutatorDefinition<PropertyDecoratorSite, Args> & { options: { validTargets: ['property'] }}): AnnotationDecorator<Args>;
+    public static define<Args extends any[]>(definition: MutatorDefinition<ParameterDecoratorSite, Args> & { options: { validTargets: ['parameter'] }}): AnnotationDecorator<Args>;
+    public static define<Args extends any[]>(definition: MutatorDefinition): AnnotationDecorator<Args>;
+    public static define(definition: MutatorDefinition<any, any[]>): AnnotationDecorator<any> {
         return this.create(definition.invoke, definition.options)();
     }
 }